ENH: Remove support for TIFF tile as 3D and dead code
This TIFF file format is a file which contains a list of directories,
if there are no directories in the file it's not a valid tiff. Code
which supported 3D tiles was conditioned on a file with no directories
and no data. Additional code was removed related to a private SGI tag
TIFFTAG_TILEDEPTH, ImageJ specific tag also pre-conditioned on this
invalid case.

Change-Id: Id4ef171c5dbe0c797be206e0059c7488719ef6e6