ENH: RGBA read images should stay unchanged.
The image that are read with the TIFFReadRGBAImage should not have
their pixel types converted. This libtiff method always read images
into an ABGR uint8 format. When the tiff file is read by this method
it should report that uint8 and RGBA is the component an pixel type
for the image. This allows the standard ITK buffer conversion to be
performed.

This change remove extra cases and template instantiations.

Additionally the RGBAImageToBuffer method was flipping the image along
the y-axis, this has been more explicitly done with the call to
TIFFReadRGBAImageOriented.

Change-Id: I022b12921d406edb09730f905213b0db811f1bd4