PERF: Refactor color table lookup
Change to use a index lookup for the color and gray scale
palettes. The remove excessive exception checking done on a per pixel
basis with the GetColor method. Additionally the InitializeColor
method was change do initialize the color palette variables.

An exception no longer occurs when an index exceeds the number of
entries in a palette. Instead a module operator is used to prevent
out of bounds memory access.

Change-Id: Ia62c65cd8b1be0d66c488540709a8fb961190cc9