COMP: Wiki examples require additional vtk modules
The Wiki examples require a VtkGlue that contains vtkTestingRendering
and vtkInterationImage but VtkGlue does not require these vtk modules.
If the Remote module WikiExamples is enabled, require the additional
modules for vtk.

NOTE: In the future we need a better way to handle itk modules that
require vtk modules.

Change-Id: I5d2c916a1c6cce95bc2971e6b9b192b04dffa27c

Conflicts:
	Modules/Bridge/VtkGlue/CMakeLists.txt