ENH: Install Python bindings into site-packages.
By installing Python bindings into the site-packages directory, instead of
trying to use a WrapITK.pth, the installation is more portable -- the absolute
paths in a WrapITK.pth do not have to be the same on the build system as on
the system the package is installed into.

The itk.py module is converted into a package to hold the swig generated
bindings.

Change-Id: I3dbbf7c569b5c5c3a8e7937b62bb661c71184f45